Addition of a toggle for effects and removal of the default check mark in the remove effects app. (#282)

+++ scripts/effects/remove-effects-handler.js
@@ -24,10 +24,16 @@ export default class RemoveEffectsHandler {
     }
 
     const selections = await this._getSelectionsFromDialog();
-    selections?.forEach(async (effectIds, actorUuid) => {
+    const {effectData,todo} = selections;
+    for (const [actorUuid,effectIds] of effectData) {
       const actor = this._foundryHelpers.getActorByUuid(actorUuid);
-      await actor.deleteEmbeddedDocuments('ActiveEffect', effectIds);
-    });
+      if (selections?.todo === 'toggle') {
+        const updates = effectIds.map((id)=>{return {_id:id,disabled:!actor.effects.get(id).disabled}});
+        await actor.updateEmbeddedDocuments('ActiveEffect',updates);
+      }
+      else if (selections?.todo === 'remove')
+        await actor.deleteEmbeddedDocuments('ActiveEffect', effectIds);
+    }
   }
 
   get _effectsByActorMappings() {
@@ -56,17 +62,23 @@ export default class RemoveEffectsHandler {
   }
 
   async _getDialog(resolve, _reject) {
+    const effectsByActorMappings = this._effectsByActorMappings;
+    for (const i of effectsByActorMappings) {
+      for (const e of i.effects) {
+        if (!!e.disabled && !e.name.includes('(Disabled')) e.name = `${e.name} (Disabled)`;
+      }
+    }
     const content = await renderTemplate(
       'modules/dfreds-convenient-effects/templates/remove-effects-dialog.hbs',
-      { effectsByActorMappings: this._effectsByActorMappings }
+      { effectsByActorMappings }
     );
     return new Dialog(
       {
-        title: 'Remove Effects',
+        title: 'Update Effects',
         content: content,
         buttons: {
           remove: {
-            icon: '<i class="fas fa-check"></i>',
+            icon: '<i class="fas fa-trash"></i>',
             label: 'Remove',
             callback: (html) => {
               const checkedData = html
@@ -89,7 +101,34 @@ export default class RemoveEffectsHandler {
                   return result;
                 }, {});
 
-              resolve(new Map(Object.entries(checkedData)));
+              resolve({effectData:new Map(Object.entries(checkedData)),todo:'remove'});
+            },
+          },
+          toggle: {
+            icon: '<i class="fas fa-toggle-off"></i>',
+            label: 'Toggle',
+            callback: (html) => {
+              const checkedData = html
+                .find('input:checked')
+                .map((idx, ele) => {
+                  const data = $(ele).data();
+                  return {
+                    actorUuid: data?.actorUuid,
+                    effectId: data?.effectId,
+                  };
+                })
+                .get()
+                .reduce((result, currentValue) => {
+                  // If an array already present for key, push it to the array. Else create an array and push the object
+                  (result[currentValue['actorUuid']] =
+                    result[currentValue['actorUuid']] || []).push(
+                    currentValue.effectId
+                  );
+                  // Return the current iteration `result` value, this will be taken as next iteration `result` value and accumulate
+                  return result;
+                }, {});
+
+              resolve({effectData:new Map(Object.entries(checkedData)),todo:'toggle'});
             },
           },
           cancel: {
@@ -102,7 +141,7 @@ export default class RemoveEffectsHandler {
         },
         default: 'cancel',
       },
-      { width: 300 }
+      { width: 300, id:'DFredsCE-removeEffects',height: "auto", resizable: true }
     );
   }
 }
